extract profile suggestions into ssb.patchwork.suggest.profile (suggest plugin)
